improve exception handling in import_osm_to_db()

+++ python/scripts/process_osm.py
@@ -120,17 +120,25 @@ def import_osm_to_db(input_file: str, force: bool, style_file_path: str = str(DE
     if not os.path.exists(style_file_path):
         raise FileNotFoundError(f"Style file {style_file_path} does not exist.")
 
-    # preprocessing
     preprocessed_file = str(RESOURCES_DIR / 'preprocessed.osm.pbf')
-    run_osmium_cmd('r', input_file, preprocessed_file)
 
-    # importing to database
-    run_osm2pgsql_cmd(CREDENTIALS, preprocessed_file, style_file_path, schema, force)
+    try:
+         # preprocessing
+        run_osmium_cmd('r', input_file, preprocessed_file)
 
-    os.remove(preprocessed_file)
+        # importing to database
+        run_osm2pgsql_cmd(CREDENTIALS, preprocessed_file, style_file_path, schema, force)
 
-    # postprocessing
-    postprocess_osm_import(CREDENTIALS, style_file_path, schema)
+        # postprocessing
+        postprocess_osm_import(CREDENTIALS, style_file_path, schema)
+    except InvalidInputError as e:
+        logger.error(f"Error during pre-processing: {e}")
+    except SubprocessError as e:
+        logger.error(f"Error during processing: {e}")
+    finally:
+        if os.path.exists(preprocessed_file):
+            os.remove(preprocessed_file)
+
 
 def parse_args(arg_list: list[str] | None) -> argparse.Namespace:
     parser = argparse.ArgumentParser(description="Process OSM files and interact with PostgreSQL database.", formatter_class=argparse.RawTextHelpFormatter)
